requirement.txt
langchain
flask
python-dotenv
openai